test: Fix simple decoding test
Problem: The unit tests for the decoder appeared to assume that
`0x00 as u16` would not deserialize into a Telemetry struct because the
Telemetry stuct contains a u32. This is incorrect because Postcard uses
varint encoding and a Telemetry struct containing only one u32 can be
deserialized from as little as a single u8. This incorrect assumption
meant that the tests were failing.

Solution: Update the tests to demonstrate that as little as 1 byte is
sufficient to deserialize a Telemetry struct.

Testing: Ran `cargo test` and saw that it succeeded.